Move example code into an example target
