+++ crates/yakui-vulkan/src/lib.rs
@@ -15,7 +15,7 @@
 //! This crate requires at least Vulkan 1.2 and a GPU with support for `VkPhysicalDeviceDescriptorIndexingFeatures.descriptorBindingPartiallyBound`.
 //! You should also, you know, enable that feature, or Vulkan Validation Layers will get mad at you. You definitely don't want that.
 //!
-//! For an example of how to use this crate, check out the cleverly named `it_works` test in `lib.rs` in the GitHub repo.
+//! For an example of how to use this crate, check out `examples/demo.rs`
 
 mod buffer;
 mod descriptors;
@@ -774,243 +774,4 @@ fn create_framebuffers(
         })
         .collect();
     framebuffers
-}
-
-#[cfg(all(windows, test))]
-mod tests {
-    use super::*;
-    use ash::vk;
-    use yakui::geometry::Vec2;
-
-    use winit::{
-        event::{ElementState, Event, KeyboardInput, VirtualKeyCode, WindowEvent},
-        event_loop::ControlFlow,
-        platform::run_return::EventLoopExtRunReturn,
-    };
-    use yakui::image;
-
-    const MONKEY_PNG: &[u8] = include_bytes!("../../bootstrap/assets/monkey.png");
-    const DOG_JPG: &[u8] = include_bytes!("../assets/dog.jpg");
-
-    #[derive(Debug, Clone)]
-    struct GuiState {
-        monkey: yakui::ManagedTextureId,
-        dog: yakui::TextureId,
-        which_image: WhichImage,
-    }
-
-    #[derive(Debug, Clone)]
-    enum WhichImage {
-        Monkey,
-        Dog,
-    }
-
-    #[test]
-    #[ignore]
-    /// Simple smoke test to make sure render screen properly pixel Vulkan.
-    fn it_works() {
-        use winit::dpi::PhysicalSize;
-
-        use crate::util::{init_winit, VulkanTest};
-
-        let (width, height) = (500, 500);
-        let (mut event_loop, window) = init_winit(width, height);
-        let mut vulkan_test = VulkanTest::new(width, height, &window);
-        let render_surface = RenderSurface {
-            resolution: vk::Extent2D { width, height },
-            format: vulkan_test.swapchain_info.surface_format.format,
-            image_views: vulkan_test.present_image_views.clone(),
-            load_op: vk::AttachmentLoadOp::CLEAR,
-        };
-        let mut yak = yakui::Yakui::new();
-        yak.set_surface_size([width as f32, height as f32].into());
-        yak.set_unscaled_viewport(yakui_core::geometry::Rect::from_pos_size(
-            Default::default(),
-            [width as f32, height as f32].into(),
-        ));
-
-        let (mut yakui_vulkan, mut gui_state) = {
-            let vulkan_context = VulkanContext::new(
-                &vulkan_test.device,
-                vulkan_test.present_queue,
-                vulkan_test.draw_command_buffer,
-                vulkan_test.command_pool,
-                vulkan_test.device_memory_properties,
-            );
-            let mut yakui_vulkan = YakuiVulkan::new(&vulkan_context, render_surface);
-            let gui_state = GuiState {
-                monkey: yak.add_texture(create_yakui_texture(
-                    MONKEY_PNG,
-                    yakui::paint::TextureFilter::Linear,
-                )),
-                dog: yakui_vulkan.create_user_texture(
-                    &vulkan_context,
-                    create_vulkan_texture_info(DOG_JPG, vk::Filter::LINEAR),
-                ),
-                which_image: WhichImage::Monkey,
-            };
-            (yakui_vulkan, gui_state)
-        };
-
-        let mut winit_initializing = true;
-
-        event_loop.run_return(|event, _, control_flow| {
-            *control_flow = ControlFlow::Poll;
-            match event {
-                Event::WindowEvent {
-                    event:
-                        WindowEvent::CloseRequested
-                        | WindowEvent::KeyboardInput {
-                            input:
-                                KeyboardInput {
-                                    state: ElementState::Pressed,
-                                    virtual_keycode: Some(VirtualKeyCode::Escape),
-                                    ..
-                                },
-                            ..
-                        },
-                    ..
-                } => *control_flow = ControlFlow::Exit,
-
-                Event::NewEvents(cause) => {
-                    if cause == winit::event::StartCause::Init {
-                        winit_initializing = true;
-                    } else {
-                        winit_initializing = false;
-                    }
-                }
-
-                Event::MainEventsCleared => {
-                    let framebuffer_index = vulkan_test.render_begin();
-                    let vulkan_context = VulkanContext::new(
-                        &vulkan_test.device,
-                        vulkan_test.present_queue,
-                        vulkan_test.draw_command_buffer,
-                        vulkan_test.command_pool,
-                        vulkan_test.device_memory_properties,
-                    );
-
-                    yak.start();
-                    gui(&gui_state);
-                    yak.finish();
-
-                    yakui_vulkan.paint(&mut yak, &vulkan_context, framebuffer_index);
-                    vulkan_test.render_end(framebuffer_index);
-                }
-                Event::WindowEvent {
-                    event: WindowEvent::Resized(size),
-                    ..
-                } => {
-                    if winit_initializing {
-                        println!("Ignoring resize during init!");
-                    } else {
-                        let PhysicalSize { width, height } = size;
-                        vulkan_test.resized(width, height);
-                        let render_surface = RenderSurface {
-                            resolution: vk::Extent2D { width, height },
-                            format: vulkan_test.swapchain_info.surface_format.format,
-                            image_views: vulkan_test.present_image_views.clone(),
-                            load_op: vk::AttachmentLoadOp::CLEAR,
-                        };
-                        yakui_vulkan.update_surface(render_surface, &vulkan_test.device);
-                        yak.set_surface_size([width as f32, height as f32].into());
-                        yak.set_unscaled_viewport(yakui_core::geometry::Rect::from_pos_size(
-                            Default::default(),
-                            [width as f32, height as f32].into(),
-                        ));
-                    }
-                }
-                Event::WindowEvent {
-                    event: WindowEvent::ScaleFactorChanged { scale_factor, .. },
-                    ..
-                } => yak.set_scale_factor(scale_factor as _),
-                Event::WindowEvent {
-                    event:
-                        WindowEvent::KeyboardInput {
-                            input:
-                                KeyboardInput {
-                                    state: ElementState::Released,
-                                    virtual_keycode,
-                                    ..
-                                },
-                            ..
-                        },
-                    ..
-                } => match virtual_keycode {
-                    Some(VirtualKeyCode::A) => {
-                        gui_state.which_image = match &gui_state.which_image {
-                            WhichImage::Monkey => WhichImage::Dog,
-                            WhichImage::Dog => WhichImage::Monkey,
-                        }
-                    }
-                    _ => {}
-                },
-                _ => (),
-            }
-        });
-
-        unsafe {
-            yakui_vulkan.cleanup(&vulkan_test.device);
-        }
-    }
-
-    fn create_vulkan_texture_info(
-        compressed_image_bytes: &[u8],
-        filter: vk::Filter,
-    ) -> VulkanTextureCreateInfo<Vec<u8>> {
-        let image = image::load_from_memory(compressed_image_bytes)
-            .unwrap()
-            .into_rgba8();
-        let resolution = vk::Extent2D {
-            width: image.width(),
-            height: image.height(),
-        };
-
-        VulkanTextureCreateInfo::new(
-            image.into_raw(),
-            vk::Format::R8G8B8A8_UNORM,
-            resolution,
-            filter,
-            filter,
-        )
-    }
-
-    fn create_yakui_texture(
-        compressed_image_bytes: &[u8],
-        filter: yakui::paint::TextureFilter,
-    ) -> yakui::paint::Texture {
-        let image = image::load_from_memory(compressed_image_bytes)
-            .unwrap()
-            .into_rgba8();
-        let size = UVec2::new(image.width(), image.height());
-
-        let mut texture = yakui::paint::Texture::new(
-            yakui::paint::TextureFormat::Rgba8Srgb,
-            size,
-            image.into_raw(),
-        );
-        texture.mag_filter = filter;
-        texture
-    }
-
-    fn gui(gui_state: &GuiState) {
-        use yakui::{column, label, row, text, widgets::Text, Color};
-        let (animal, texture): (&'static str, yakui::TextureId) = match gui_state.which_image {
-            WhichImage::Monkey => ("monkye", gui_state.monkey.into()),
-            WhichImage::Dog => ("dog haha good boy", gui_state.dog.into()),
-        };
-        column(|| {
-            row(|| {
-                label("Hello, world!");
-
-                let mut text = Text::new(48.0, "colored text!");
-                text.style.color = Color::RED;
-                text.show();
-            });
-
-            text(96.0, format!("look it is a {animal}"));
-
-            image(texture, Vec2::new(400.0, 400.0));
-        });
-    }
 }
